# openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Data Ingestion & Privacy Intelligence API",
    "version": "1.0.0",
    "description": "Proteccio Discover backend covering Week 1 ingestion, Week 2 discovery/classification/mapping/profiling, and Week 3 governance/compliance intelligence: privacy risk, compliance controls, remediation workflows, dashboards, reporting, RBAC, alerting, and global search.\n\n**Auth:** When `JWT_SECRET` is set, use `POST /api/auth/login` and send `Authorization: Bearer <JWT>`. When only `API_KEY` is set, send `Authorization: Bearer <API_KEY>` or `X-API-Key: <API_KEY>` on `/api/*` except `/health` and documentation routes.\n\n**Docs UI:** `GET /docs` (Swagger UI). **Raw spec:** `GET /openapi.json`."
  },
  "servers": [
    {
      "url": "{baseUrl}",
      "variables": {
        "baseUrl": {
          "default": "http://localhost:3000"
        }
      }
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKeyHeader": []
    }
  ],
  "paths": {
...
    "/api/search/datasets": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "Search governance datasets (compliance, risk, remediation, multi-label, keyword, cursor)",
        "parameters": [
          {
            "name": "riskLevel",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "low",
                "medium",
                "high",
                "critical"
              ]
            }
          },
          {
            "name": "riskLevels",
            "in": "query",
            "schema": {
              "type": "string",
              "description": "Comma-separated risk levels"
            }
          },
          {
            "name": "minRiskScore",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "maxRiskScore",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "classification",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "classifications",
            "in": "query",
            "schema": {
              "type": "string",
              "description": "Comma-separated labels (AND by default)"
            }
          },
          {
            "name": "classificationMode",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "and",
                "or"
              ]
            }
          },
          {
            "name": "sourceType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sourceName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "systemId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "datasetId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "detectionType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "detectionCategory",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "detectionCategories",
            "in": "query",
            "schema": {
              "type": "string",
              "description": "e.g. aadhaar,pan (AND)"
            }
          },
          {
            "name": "detectionMode",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "and",
                "or"
              ]
            }
          },
          {
            "name": "complianceRegulation",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "GDPR",
                "DPDP",
                "HIPAA",
                "CCPA",
                "ISO27001"
              ]
            }
          },
          {
            "name": "complianceViolation",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "complianceStatus",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hasUnresolvedRemediation",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string",
              "description": "Keyword search"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "updatedAt",
                "riskScore",
                "riskLevel",
                "sourceName",
                "complianceScore"
              ]
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mappedOnly",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "page of results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/search/global": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "Global keyword search across datasets, fields, remediation, lineage, sources",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "types",
            "in": "query",
            "schema": {
              "type": "string",
              "description": "Comma-separated: datasets,fields,remediation,lineage,sources"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "grouped search hits",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/search/lineage": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "Search data lineage by dataset, system, flow, or related source name",
        "parameters": [
          {
            "name": "datasetId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "systemId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "direction",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "upstream",
                "downstream",
                "both"
              ]
            }
          },
          {
            "name": "flowKind",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "relatedSourceName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "lineage hits",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/search/remediation": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "Search remediation tickets (unresolved filter, keyword, cursor)",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "unresolved",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "datasetId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "remediation page",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/search/mapped-fields": {
      "get": {
        "tags": [
          "Search"
        ],
        "summary": "Search mapped fields",
        "parameters": [
          {
            "name": "datasetId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sensitiveCategory",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "page",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "JWT from POST /api/auth/login when JWT_SECRET is set; otherwise API_KEY value."
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Same value as API_KEY when set."
      }
    },
...
  }
}